pyproject.toml
[tool.poetry]
name = "api-carbon-monitor"
version = "0.1.0"
description = "API consumption manager that also tracks your carbon, electricity, and water usage oPI dependencies"
authors = ["Dillon O'Leary <doleary@wisc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.59.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
